Choice between several server types. Launching script improvements.

+++ RESTNavServer/launchers/web/scratch/fax-processor/fax.processor.sh
@@ -74,12 +74,24 @@ then
   # See https://rawsec.ml/en/python-3-simplehttpserver/
   #
   echo -e "Starting python server"
+  PYTHON_3=$(which python3)
+  if [[ "${PYTHON_3}" == "" ]]
+  then
+    echo -e "python3 must be installed on your system... Exiting."
+    exit 1
+  fi
   python3 -m http.server ${HTTP_PORT} &
   SERVER_PROCESS_ID=$(echo $!)
   echo -e "To kill the server, used PID ${SERVER_PROCESS_ID}"
 elif [[ "${SERVER_FLAVOR}" == "node" ]]
 then
   echo -e "Starting node server"
+  NODE_JS=$(which node)
+  if [[ "${NODE_JS}" == "" ]]
+  then
+    echo -e "node must be installed on your system... Exiting."
+    exit 1
+  fi
   node server.js --verbose:${VERBOSE} --port:${HTTP_PORT} &
   SERVER_PROCESS_ID=$(echo $!)
   echo -e "To kill the server, used PID ${SERVER_PROCESS_ID}"
@@ -87,9 +99,17 @@ elif [[ "${SERVER_FLAVOR}" == "java" ]]
 then
   echo -e "Starting java server"
   # CP=$(find ~/repos/raspberry-coffee -name http-tiny-server-1.0-all.jar)
-  CP=../../../../../http-tiny-server/build/libs/http-tiny-server-1.0-all.jar
+  JAR_FILE=../../../../../http-tiny-server/build/libs/http-tiny-server-1.0-all.jar
+  if [[ ! -f ${JAR_FILE} ]]
+  then
+    echo -e "${JAR_FILE} not found where expected. Exiting"
+    exit 1
+  fi
+  CP=${JAR_FILE}
   JAVA_OPTIONS="${JAVA_OPTIONS} -Dhttp.verbose=${VERBOSE}"
   JAVA_OPTIONS="${JAVA_OPTIONS} -Dhttp.port=${HTTP_PORT}"
+  JAVA_OPTIONS="${JAVA_OPTIONS} -Dstatic.docs=/"
+  JAVA_OPTIONS="${JAVA_OPTIONS} -Dhttp.super.verbose=true"
   echo -e "Will run: java -cp ${CP} ${JAVA_OPTIONS} http.HTTPServer &"
   java -cp ${CP} ${JAVA_OPTIONS} http.HTTPServer &
   SERVER_PROCESS_ID=$(echo $!)
